refactor: 추상 객체가 Random에만 국한되지 않도록 클래스명 변경
