+++ src/main/java/study/racing/domain/NumberGenerator.java
@@ -1,0 +1,6 @@
+package study.racing.domain;
+
+public interface NumberGenerator {
+
+    int random(int bound);
+}
